move loglevel to package logging
