+++ client/src/commonMain/kotlin/com/algolia/search/client/ClientRecommend.kt
@@ -6,7 +6,7 @@ import com.algolia.search.client.internal.ClientRecommendImpl
 import com.algolia.search.configuration.Configuration
 import com.algolia.search.configuration.ConfigurationRecommend
 import com.algolia.search.configuration.Credentials
-import com.algolia.search.configuration.LogLevel
+import com.algolia.search.logging.LogLevel
 import com.algolia.search.configuration.internal.Credentials
 import com.algolia.search.configuration.internal.DEFAULT_LOG_LEVEL
 import com.algolia.search.endpoint.EndpointRecommend
