+++ client/src/commonMain/kotlin/com/algolia/search/configuration/ConfigurationAnalytics.kt
@@ -7,6 +7,7 @@ import com.algolia.search.configuration.internal.ConfigurationAnalyticsImpl
 import com.algolia.search.configuration.internal.DEFAULT_LOG_LEVEL
 import com.algolia.search.configuration.internal.DEFAULT_READ_TIMEOUT
 import com.algolia.search.configuration.internal.DEFAULT_WRITE_TIMEOUT
+import com.algolia.search.logging.LogLevel
 import com.algolia.search.model.APIKey
 import com.algolia.search.model.ApplicationID
 import com.algolia.search.transport.internal.hosts
