+++ client/src/commonMain/kotlin/com/algolia/search/configuration/internal/ConfigurationPlacesImpl.kt
@@ -2,7 +2,7 @@ package com.algolia.search.configuration.internal
 
 import com.algolia.search.configuration.Compression
 import com.algolia.search.configuration.ConfigurationPlaces
-import com.algolia.search.configuration.LogLevel
+import com.algolia.search.logging.LogLevel
 import com.algolia.search.configuration.RetryableHost
 import com.algolia.search.configuration.internal.extension.getHttpClient
 import io.ktor.client.HttpClient
