+++ client/src/commonMain/kotlin/com/algolia/search/configuration/internal/ConfigurationRecommendImpl.kt
@@ -3,7 +3,7 @@ package com.algolia.search.configuration.internal
 import com.algolia.search.client.ClientRecommend
 import com.algolia.search.configuration.Compression
 import com.algolia.search.configuration.ConfigurationRecommend
-import com.algolia.search.configuration.LogLevel
+import com.algolia.search.logging.LogLevel
 import com.algolia.search.configuration.RetryableHost
 import com.algolia.search.configuration.internal.extension.getHttpClient
 import com.algolia.search.model.APIKey
