+++ client/src/commonMain/kotlin/com/algolia/search/configuration/internal/Constants.kt
@@ -1,6 +1,6 @@
 package com.algolia.search.configuration.internal
 
-import com.algolia.search.configuration.LogLevel
+import com.algolia.search.logging.LogLevel
 
 internal const val DEFAULT_WRITE_TIMEOUT: Long = 30000L
 internal const val DEFAULT_READ_TIMEOUT: Long = 5000L
