+++ client/src/commonMain/kotlin/com/algolia/search/logging/LogLevel.kt
@@ -1,0 +1,8 @@
+package com.algolia.search.logging
+
+/**
+ * Http client logging log level.
+ */
+public enum class LogLevel {
+    All, Headers, Body, Info, None
+}
